fix トップ画像取得したまで。
